server/requirements.txt
fastapi[standard]
uvicorn
python-dotenv
sqlalchemy
pydantic_settings
psycopg2-binary
passlib[bcrypt]
bcrypt<4.0.0
python-jose[cryptography]
